+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 ## in develop
 
 * Handles empty scatters in WDL workflows
+* Where applet calls executables, adds executables to bundledDepends to support cloning of workflows
 * Enables other metadata (title, description, version, etc) to be set via extras.json
 * Creates hard- rather than soft-links in the input directory, so that linked files are accessible from within containers